Allow multiple named Kafka consumers to exist.
The consumers are referenced by name argument that is added to each consumer keyword.
The default client name is 'default' if no name argument is provided.

A consumer dict is kept in the Consumer object. The dictionary stores customer instances
by name. An instance is removed from the dictionary when `Close` keyword is called.